[doc] fix online doco links
Some links in README etc were using the old `doc-project/` directory name, which was merged in to `docs/` so it could be used with GitHub Pages publishing from the master branch. And a couple of the links were missing the `user-guide/` directory element under that.

Also tweaked the Markdown formatting in Design-and-Justification to make markdownlint happy.