[code] combine some +internal dirs in a single +tblish, for easy browsing and scope clarity
Except for the +tblish/+examples +internal dir, which is intended to be internal to specifically the examples, and not Tablicious as a whole.

IMHO, this consolidation makes it more clear that these +internal subpackages are intended for the internal use of the Tablicious package (in the sense of the whole library or OF package, not a +blah namespace package) as a whole, and not the specific namespace packages inside Tablicious. It also collects them in one spot in the dir tree for easier browsing, and may make searching the code base (e.g. with grep) easier.