[test] refine the runtests.sh script for output capture and running from any directory
This makes it diverge from the "standard" runtests.sh script that I think is found in other Octave Forge packages. I'm fine with that; this script can be Tablicious-specific. (It's in the nonstandard dev-tools/ subdir, so I don't think any standard Octave tools like pkg could be depending on it.)