table: deprecate nonstandard rows() and columns() methods, replaced by height() and width()
The rows() and columns() methods were added quite a while ago, I think before height() and
width() were standardized. The height() and width() functions fully replace them, so
we'll remove them to keep the interface simpler and more consistent with Matlab's table
interface. But give users a deprecation-with-warning period so they don't face a hard
breaking change in a single release. I'm targeting this deprecation for Tablicious 0.4.0,
and the removal for 0.5.0.