table: remove reshape(), resize(), and vec() methods
Reshape in the Matlab table API is a method on tabular, but it just raises an error so it's not useful and there's no actual functionality to implement here. Seems like Tablicious's users would prefer to see the method absent from table's interface here, and let whatever error or other behavior Octave's built-in reshape() does happen, even if that's less Matlab-conformant. (The Statistics package in particular would like to see a smaller table interface.)

The vec method is an Octave extension, not in Matlab, so I'm fine with not supporting that. Without a reshape method override, in Octave 8.x, calling vec() raises the error "error: table.subsref: ()-indexing of table requires exactly two arguments", which makes sense if vec() is defined in terms of ()-indexing with a single argument. Seems that users won't be passing table arrays as a drop-in replacement for other types of vectors, so having `vec(tbl)` behavior that is a successful identity operation for tables which are already column vectors (1-var tables) may be less important than reducing the interface size to make users comfortable.

Resize does need a specialization method at this point. The core Octave resize() function in 8.x returns "error: octave_base_value::resize (): wrong type argument 'object'". But in the Matlab table interface, resize() is just a function with no table method override. And I don't have a useful implementation here yet. So may as well remove it, especially if Statistics wants to supply one.