[code] move main classes under @<class> dirs, for code browsing purposes
This change is purely for repo organization purposes. It groups the main classes up at the top of a directory listing in browsers like text editors, either because they're directories, or because of the leading "@". Shouldn't change any code behavior. In the future, this will allow for adding class-specific resource/ subdirs, but that's not the reason for doing this now.

Doing this change required getting all the functions out of the inst/private directory and in to +tblish/+internal, so they could be called from anywhere and shared between source code in different directories.