fix syntax(-ish) error (bad "foo (...)" spaces inside "[...]") in datetime, tblish.internal.mycombvec, and others
This was raised by Markus Mützel for this particular case in mycombvec, and then I checked for others with grep like the following, which gets false positives (esp for the "{...}" case), but will catch a lot of actual issues like this.

    grep -rn '\[.*[A-Za-z0-9_] (.*\]' inst
    grep -rn '^[^#]*{.*[A-Za-z0-9_] (.*}' inst

I think (but did not verify through testing) that the impact was:

* The calendarDuration ones broke its proxyKeys() and sort().
* The mycombvec() one broke it entirely, which broke a couple pretty-printing display functions.
* The localdate ones broke its sort() and unique().
* The datetime ones broke its sort() and unique().
* The splitapply_impl and mustBeCellstr ones broke those functions.

Looks like none of these were covered by tests.

I bet these were introduced by the recent "conver to GNU Octave code style" change I did, where I added spaces to function calls, like "foo ()" instead of "foo()". Was probably overzealous and added spaces in some places where that syntax is *not* valid, inside "[...]" and "{...}" concatenation.