fix size-vs-numel bug in duration ctor that broke conversion of cellstr row vectors
The `for i = 1:size (strs)` is a bug; it should be `numel (strs)`. When using `size()`, you get a vector, then I think the `1:x` colon operator just uses the first element of the RHS, so it under-counts if dim 2 or higher is nonzero, and leaves the ignored elements as NaN in the output.

See: https://github.com/apjanke/octave-tablicious/issues/134