datetime: colon: fix hi bound test, use multiplied inc instead of repeatedly added inc for caldurs
Addresses https://github.com/apjanke/octave-tablicious/issues/135.

The hi bound test was a dumb off-by-one-ish logic bug in the control flow logic.

Changing the calendarDuration inc behavior to multiply the inc on each loop and add it to the original lo, instead of repeatedly adding the inc to a working value, handles "displacement" due to variable-length months and the like, where the working value can get jiggered around by month-end normalization (capping to last day of month) and similar stuff.

This doesn't fix the actual month-end capping/normalization for calendarDuration arithmetic, but makes colon support it when it does get fixed, over in datetime.plus or calendarDuration.plus. That fix should be coming soon.