[time] restructure calendarDuration, removing Sign, Years, and IsNan fields
Years are just defined to be 12-month units, so they can be expressed as months. The NaN-ness can be carried in the other numeric fields. And different components can have different signs, so a separate Sign field is kind of ill-defined, and redundant.

Removes the documentation for properties that weren't public in the first place.