fix string.regexp breakage due to demote_strings not working on non-string inputs
The move of demote_strings to be an object method broke its use for varargins that contain no string objects. That's the case with regexprep and maybe a couple other methods, which peel off their leading arguments to named variables instead of grabbing the whole arg list to varargin (for signature readability).